+++ src/propulsion/rpm_regulator.cpp
@@ -21,8 +21,8 @@ uint32_t RpmRegulator::calculateRpm(const data::nav_t actual_velocity, const int
 uint32_t RpmRegulator::calculateOptimalRpm(const data::nav_t actual_velocity)
 {
   // polynomial values from simulation
-  return std::round(0.32047 * actual_velocity * actual_velocity + 297.72578 * actual_velocity
-                    + 1024.30824);
+  return static_cast<uint32_t>(std::round(0.32047 * actual_velocity * actual_velocity
+                                          + 297.72578 * actual_velocity + 1024.30824));
 }
 
 uint32_t RpmRegulator::step(const int32_t optimal_rpm, const int32_t actual_rpm)
